+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/modules/oppgavebekreftelse/Oppgavebekreftelse.tsx
@@ -3,9 +3,9 @@ import { getOppgaveStatusText, getOppgaveTittel } from '@innsyn/utils/textUtils'
 import { Heading, VStack } from '@navikt/ds-react';
 import { useAppIntl } from '@shared/i18n';
 import { BekreftelseOppgave } from '@shared/types/Oppgave';
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 
-import { OppgavebekreftelseContext, useOppgavebekreftelse } from './hooks/useOppgavebekreftelse';
+import { OppgavebekreftelseContext } from './hooks/useOppgavebekreftelse';
 import { Besvart, Kvittering, Ubesvart } from './OppgavebekreftelseParts';
 
 interface Props {
@@ -19,14 +19,18 @@ const Oppgavebekreftelse = ({ oppgave, deltakerNavn, children, initialVisKvitter
     const appIntl = useAppIntl();
     const [visKvittering, setVisKvittering] = useState(initialVisKvittering);
 
+    const contextValue = useMemo(
+        () => ({
+            oppgave,
+            deltakerNavn,
+            visKvittering,
+            setVisKvittering,
+        }),
+        [oppgave, deltakerNavn, visKvittering, setVisKvittering],
+    );
+
     return (
-        <OppgavebekreftelseContext.Provider
-            value={{
-                oppgave,
-                deltakerNavn,
-                visKvittering,
-                setVisKvittering,
-            }}>
+        <OppgavebekreftelseContext.Provider value={contextValue}>
             <VStack gap="6">
                 <div>
                     <OppgaveStatusTag
@@ -47,5 +51,4 @@ Oppgavebekreftelse.Ubesvart = Ubesvart;
 Oppgavebekreftelse.Besvart = Besvart;
 Oppgavebekreftelse.Kvittering = Kvittering;
 
-export { Oppgavebekreftelse, useOppgavebekreftelse };
 export default Oppgavebekreftelse;