+++ apps/ungdomsytelse-deltaker/src/apps/søknad/api/deltakelse/markerDeltakelseSomSøkt.ts
@@ -4,7 +4,6 @@ import { Deltakelse } from '@navikt/ung-deltakelse-opplyser-api-deltaker';
 export const markerDeltakelseSomSøkt = async (deltakelseId: string): Promise<void> => {
     try {
         await Deltakelse.markerDeltakelseSomSøkt({ path: { id: deltakelseId }, headers: commonRequestHeader });
-        return Promise.resolve();
     } catch (e) {
         throw handleApiError(e, 'markerDeltakelseSomSøkt');
     }