+++ packages/k9-sak-innsyn-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.1.0
+
+### Minor Changes
+
+- Codegen
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.0.36
 
 ### Patch Changes